drm/xe/kunit: Simplify xe_bo live tests code layout
BugLink: https://bugs.launchpad.net/bugs/2091444

The test case logic is implemented by the functions compiled as
part of the core Xe driver module and then exported to build and
register the test suite in the live test module.

But we don't need to export individual test case functions, we may
just export the entire test suite. And we don't need to register
this test suite in a separate file, it can be done in the main
file of the live test module.

Signed-off-by: Michal Wajdeczko <michal.wajdeczko@intel.com>
Reviewed-by: Lucas De Marchi <lucas.demarchi@intel.com>
Link: https://patchwork.freedesktop.org/patch/msgid/20240708111210.1154-2-michal.wajdeczko@intel.com
(cherry picked from commit d6e850acc716d0fad756f09488d198db2077141e)
Signed-off-by: Thibault Ferrante <thibault.ferrante@canonical.com>