drm/amd/display: Fetch the EDID from _DDC if available for eDP
BugLink: https://bugs.launchpad.net/bugs/2111545

Some manufacturers have intentionally put an EDID that differs from
the EDID on the internal panel on laptops.

Attempt to fetch this EDID if it exists and prefer it over the EDID
that is provided by the panel. If a user prefers to use the EDID from
the panel, offer a DC debugging parameter that would disable this.

Signed-off-by: Mario Limonciello <mario.limonciello@amd.com>
Reviewed-by: Alex Hung <alex.hung@amd.com>
Tested-by: Daniel Wheeler <daniel.wheeler@amd.com>
Signed-off-by: Rodrigo Siqueira <Rodrigo.Siqueira@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>
(backported from commit c6a837088bedef74e38f47eb65247dea1d835634)
[AaronMa: context change and ignore one space change of non-existent
function.]
Signed-off-by: Aaron Ma <aaron.ma@canonical.com>
Acked-by: Stefan Bader <stefan.bader@canonical.com>
Acked-by: Kuba Pawlak <kuba.pawlak@canonical.com>
Signed-off-by: Mehmet Basaran <mehmet.basaran@canonical.com>